+++ src/clock.c
@@ -1,7 +1,7 @@
 #include <clock.h>
 
 
-u128 uptime = 0;
+u64 uptime = 0;
 
 void clock_ihc_handler(u8 n) {
     uptime++;
